% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baconipsum.R
\name{baconipsum}
\alias{baconipsum}
\title{Obtain meaty placeholder text using the Bacon Ipsum API}
\source{
\url{https://baconipsum.com/}
}
\usage{
baconipsum(
  n_paras = 5L,
  n_sentences = NULL,
  type = c("all-meat", "meat-and-filler"),
  start_with_lorem = TRUE
)
}
\arguments{
\item{n_paras}{Integer. The number of paragraphs to generate. Defaults to 5.}

\item{n_sentences}{Integer, optional. The number of sentences to generate.
Overrides \code{n_paras}.}

\item{type}{Character. One of "all-meat" for meat words only, or
"meat-and-filler" for meat mixed with random lorem ipsum words. Defaults to
"all-meat".}

\item{start_with_lorem}{Logical. If \code{TRUE}, starts the first paragraph with 'Bacon ipsum dolor sit amet'. Defaults to \code{TRUE}.}
}
\description{
Generates placeholder text that, on first glance, looks like a regular Lorem
Ipsum, but is actually meat-related words.
}
\seealso{
Other whimsical generators: 
\code{\link{litipsum}()}
}
\concept{whimsical generators}
