% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corporatelorem.R
\name{corporatelorem}
\alias{corporatelorem}
\title{Generate Corporate-sounding placeholder text using the CorporateLorem API}
\source{
\url{https://corporatelorem.kovah.de/}
}
\usage{
corporatelorem(n = 3, paragraphs = TRUE, textformat = TRUE)
}
\arguments{
\item{n}{Integer. The number of paragraphs to generate. Defaults to 3.}

\item{paragraphs}{Logical. If \code{TRUE}, shows paragraph tags in the output. Defaults to \code{TRUE}.}

\item{textformat}{Logical. If \code{TRUE}, returns a single text string. Otherwise returns a named list of source and paragraphs.. Defaults to \code{TRUE}.}
}
\description{
Generates random "corporate-style" placeholder text using the CorporateLorem
API. The text is English-language and is partially adapted from articles by
the NY Times, The Guardian and The Verge.
}
\seealso{
Other practical generators: 
\code{\link{loripsum}()}
}
\concept{practical generators}
