% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devlorem.R
\name{devlorem}
\alias{devlorem}
\title{Generate Paragraphs of Text impersonating famous actors}
\source{
\url{https://devlorem.kovah.de/}
}
\usage{
devlorem(n = 5L, paragraphs = TRUE)
}
\arguments{
\item{n}{Integer. The number of paragraphs to generate. Defaults to 5.}

\item{paragraphs}{Logical. If \code{TRUE}, shows paragraph tags in the output. Defaults to \code{TRUE}.}
}
\description{
Generates random text impersonating famous actors, like Daisy Ridley and
Morgan Freeman.
}
\seealso{
Other whimsical generators: 
\code{\link{baconipsum}()},
\code{\link{hipsteripsum}()},
\code{\link{litipsum}()},
\code{\link{loremricksum}()},
\code{\link{metaphorpsum}()}
}
\concept{whimsical generators}
