% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinoipsum.R
\name{dinoipsum}
\alias{dinoipsum}
\title{Randomly Generate Dinosaur Names using the Dinoipsum API}
\source{
\url{https://dinoipsum.com/}
}
\usage{
dinoipsum(n_paras = 10L, n_words = 30L)
}
\arguments{
\item{n_paras}{Integer. The number of paragraphs to generate. Defaults to 10.}

\item{n_words}{Integer. The number of words to generate per paragraph.
Defaults to 30.}
}
\description{
Generates paragraphs of dinosaur names using the
}
\seealso{
Other whimsical generators: 
\code{\link{baconipsum}()},
\code{\link{devlorem}()},
\code{\link{hipsteripsum}()},
\code{\link{litipsum}()},
\code{\link{loremricksum}()},
\code{\link{metaphorpsum}()}
}
\concept{whimsical generators}
