% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hipsteripsum.R
\name{hipsteripsum}
\alias{hipsteripsum}
\title{Obtain hipster placeholder text using the Hipster Ipsum API}
\source{
\url{https://hipsum.co/}
}
\usage{
hipsteripsum(
  n_paras = 5L,
  n_sentences = NULL,
  type = c("hipster-centric", "hipster-latin"),
  start_with_lorem = TRUE
)
}
\arguments{
\item{n_paras}{Integer. The number of paragraphs to generate. Defaults to 5.}

\item{n_sentences}{Integer, optional. The number of sentences to generate.
Overrides \code{n_paras}.}

\item{type}{Character. One of "hipster-centric" for hipster words only, or
"hipster-latin" for hipster mixed with random lorem ipsum words. Defaults to
"hipster-centric".}

\item{start_with_lorem}{Logical. If \code{TRUE}, starts the first paragraph with 'I'm baby'. Defaults to \code{TRUE}.}
}
\description{
Generates pretentious placeholder text. You wouldn't understand, man!
}
\seealso{
Other whimsical generators: 
\code{\link{baconipsum}()},
\code{\link{litipsum}()}
}
\concept{whimsical generators}
