% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipsum.R
\name{lipsum}
\alias{lipsum}
\title{Lorem Ipsum}
\usage{
lipsum(n = 1)
}
\arguments{
\item{n}{The number of lorem ipsum paragraphs you want.}
}
\description{
This function will print the 'lorem ipsum' placeholder in your console, saving you the hassle of 3 seconds of searching for it and breaking your workflow.
}
\examples{
lipsum(n = 1)
}
