% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipsum_actor.R
\name{lipsum_actor}
\alias{lipsum_actor}
\title{Generate Paragraphs of Text impersonating famous actors}
\source{
\url{https://devlorem.kovah.de/}
}
\usage{
lipsum_actor(n = 5L, paragraphs = TRUE)
}
\arguments{
\item{n}{Integer. The number of paragraphs to generate. Defaults to 5.}

\item{paragraphs}{Logical. If \code{TRUE}, shows paragraph tags in the output. Defaults to \code{TRUE}.}
}
\description{
Generates random text impersonating famous actors, like Daisy Ridley and
Morgan Freeman.
}
\seealso{
Other whimsical generators: 
\code{\link{lipsum_bacon}()},
\code{\link{lipsum_dino}()},
\code{\link{lipsum_hipster}()},
\code{\link{lipsum_literature}()},
\code{\link{lipsum_metaphor}()},
\code{\link{lipsum_rick}()},
\code{\link{lipsum_starwars}()}
}
\concept{whimsical generators}
