% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipsum_dino.R
\name{lipsum_dino}
\alias{lipsum_dino}
\title{Randomly Generate Dinosaur Names using the Dinoipsum API}
\source{
\url{https://dinoipsum.com/}
}
\usage{
lipsum_dino(n_paras = 10L, n_words = 30L)
}
\arguments{
\item{n_paras}{Integer. The number of paragraphs to generate. Defaults to 10.}

\item{n_words}{Integer. The number of words to generate per paragraph.
Defaults to 30.}
}
\description{
Generates paragraphs of dinosaur names using the
}
\seealso{
Other whimsical generators: 
\code{\link{lipsum_actor}()},
\code{\link{lipsum_bacon}()},
\code{\link{lipsum_hipster}()},
\code{\link{lipsum_literature}()},
\code{\link{lipsum_metaphor}()},
\code{\link{lipsum_rick}()}
}
\concept{whimsical generators}
