% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipsum_literature.R
\name{lipsum_literature}
\alias{lipsum_literature}
\title{Obtain a random passage from copyright-free literature using the Lit Ipsum
API}
\source{
\url{https://litipsum.com/}
}
\usage{
lipsum_literature(n = NULL, tag = NULL, source = NULL)
}
\arguments{
\item{n}{Integer, optional. The number of paragraphs to generate. The
default, \code{NULL}, generates one "page" of text from the \code{source}.}

\item{tag}{Character, optional. One of "p" or "li", which renders the
paragraphs with p or li tags, respectively.}

\item{source}{Character, optional. The book from which to take the
placeholder text. One of adventures-sherlock-holmes, dr-jekyll-and-mr-hyde, dracula, evelina, life-of-samuel-johnson, picture-of-dorian-gray, pride-and-prejudice. The default,
\code{NULL}, chooses one at random.}
}
\description{
Choose passages from copyright-free literature courtesy of the \href{http://www.gutenberg.org/}{Gutenberg Project}.
}
\seealso{
Other whimsical generators: 
\code{\link{lipsum_actor}()},
\code{\link{lipsum_bacon}()},
\code{\link{lipsum_dino}()},
\code{\link{lipsum_hipster}()},
\code{\link{lipsum_metaphor}()},
\code{\link{lipsum_rick}()}
}
\concept{whimsical generators}
