% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipsum_markdown.R
\name{lipsum_markdown}
\alias{lipsum_markdown}
\title{Generate markdown-formatted Lorem Ipsum using the Lorem Markdownum API}
\source{
\url{https://jaspervdj.be/lorem-markdownum/}
}
\usage{
lipsum_markdown(
  n = 5L,
  no_headers = FALSE,
  no_code = FALSE,
  no_inline_markup = FALSE,
  no_quotes = FALSE,
  no_lists = FALSE,
  no_external_links = FALSE,
  no_wrapping = FALSE,
  underline_headers = FALSE,
  reference_links = FALSE,
  underscore_em = FALSE,
  underscore_strong = FALSE,
  fenced_code_blocks = FALSE
)
}
\arguments{
\item{n}{Integer. The number of "blocks" to generate. Defaults to 5.}

\item{no_headers}{Logical. If \code{TRUE}, removes headers. Defaults to \code{FALSE}.}

\item{no_code}{Logical. If \code{TRUE}, removes code snippets. Defaults to \code{FALSE}.}

\item{no_inline_markup}{Logical. If \code{TRUE}, removes inline markup. Defaults to \code{FALSE}.}

\item{no_quotes}{Logical. If \code{TRUE}, removes blockquotes. Defaults to \code{FALSE}.}

\item{no_lists}{Logical. If \code{TRUE}, removes lists. Defaults to \code{FALSE}.}

\item{no_external_links}{Logical. If \code{TRUE}, replaces external links with internal links. Defaults to \code{FALSE}.}

\item{no_wrapping}{Logical. If \code{TRUE}, prevents text wrapping. Defaults to \code{FALSE}.}

\item{underline_headers}{Logical. If \code{TRUE}, headers are underlined. Defaults to \code{FALSE}.}

\item{reference_links}{Logical. If \code{TRUE}, uses reference-style links. Defaults to \code{FALSE}.}

\item{underscore_em}{Logical. If \code{TRUE}, styles \if{html}{\out{<em>}} using underscores. Defaults to \code{FALSE}.}

\item{underscore_strong}{Logical. If \code{TRUE}, styles \if{html}{\out{<strong>}} using double underscores. Defaults to \code{FALSE}.}

\item{fenced_code_blocks}{Logical. If \code{TRUE}, fences code blocks with back-ticks. Defaults to \code{FALSE}.}
}
\description{
Generates customisable placeholder text using the Lorem Markdownum API, formatted for markdown.
}
\seealso{
Other practical generators: 
\code{\link{lipsum_corporate}()},
\code{\link{lipsum_loripsum}()},
\code{\link{lipsum_seazon}()}
}
\concept{practical generators}
