% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipsum_metaphor.R
\name{lipsum_metaphor}
\alias{lipsum_metaphor}
\title{Obtain wise, pedantic metaphors using the Metaphorpsum API}
\source{
\url{http://metaphorpsum.com/}
}
\usage{
lipsum_metaphor(n_paras = 2L, n_sentences = 4L, p = FALSE)
}
\arguments{
\item{n_paras}{Integer. The number of paragraphs to generate. Defaults to 2.}

\item{n_sentences}{Integer. The number of sentences to generate per
paragraph. Defaults to 4.}

\item{p}{Logical. If \code{TRUE}, wraps paragraphs in paragraph tags. Defaults to \code{FALSE}.}
}
\description{
Generates place holder text of nonsensical metaphors.
}
\seealso{
Other whimsical generators: 
\code{\link{lipsum_actor}()},
\code{\link{lipsum_bacon}()},
\code{\link{lipsum_dino}()},
\code{\link{lipsum_hipster}()},
\code{\link{lipsum_literature}()},
\code{\link{lipsum_rick}()}
}
\concept{whimsical generators}
