% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipsum_rick.R
\name{lipsum_rick}
\alias{lipsum_rick}
\title{Randomly Generate Rick & Morty Quotes}
\source{
\url{http://loremricksum.com/}
}
\usage{
lipsum_rick(n_paras = 2L, n_sentences = 4L)
}
\arguments{
\item{n_paras}{Integer. The number of paragraphs to generate. Defaults to 2.}

\item{n_sentences}{Integer. The number of sentences to generate per
paragraph. Defaults to 4.}
}
\description{
Generates paragraphs of quotes from Rick and Morty. Note that the output may
contain obscenities.
}
\seealso{
Other whimsical generators: 
\code{\link{lipsum_actor}()},
\code{\link{lipsum_bacon}()},
\code{\link{lipsum_dino}()},
\code{\link{lipsum_hipster}()},
\code{\link{lipsum_literature}()},
\code{\link{lipsum_metaphor}()}
}
\concept{whimsical generators}
