% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipsum_starwars.R
\name{lipsum_starwars}
\alias{lipsum_starwars}
\title{Generate Paragraphs of Dialogue from Star Wars films}
\source{
\url{https://forcemipsum.com/api}
}
\usage{
lipsum_starwars(n = 10L, content = "episode-4")
}
\arguments{
\item{n}{Integer. The number of paragraphs/names to generate. Defaults to 10.}

\item{content}{Character. Either \code{"episode-x"} (where \code{x} is 1-9),
\code{"characters"} or \code{"places"}. Defaults to \code{"episode-4"}.}
}
\description{
Generates random snippets of dialogue from Star Wars films, or lists of
names/places from the Star Wars universe.
}
\seealso{
Other whimsical generators: 
\code{\link{lipsum_actor}()},
\code{\link{lipsum_bacon}()},
\code{\link{lipsum_dino}()},
\code{\link{lipsum_hipster}()},
\code{\link{lipsum_literature}()},
\code{\link{lipsum_metaphor}()},
\code{\link{lipsum_rick}()}
}
\concept{whimsical generators}
