% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loremricksum.R
\name{loremricksum}
\alias{loremricksum}
\title{Randomly Generate Rick & Morty Quotes}
\source{
\url{http://loremricksum.com/}
}
\usage{
loremricksum(n_paras = 2L, n_sentences = 4L)
}
\arguments{
\item{n_paras}{Integer. The number of paragraphs to generate. Defaults to 2.}

\item{n_sentences}{Integer. The number of sentences to generate per
paragraph. Defaults to 4.}
}
\description{
Generates paragraphs of quotes from Rick and Morty. Note that the output may
contain obscenities.
}
\seealso{
Other whimsical generators: 
\code{\link{baconipsum}()},
\code{\link{devlorem}()},
\code{\link{dinoipsum}()},
\code{\link{hipsteripsum}()},
\code{\link{litipsum}()},
\code{\link{metaphorpsum}()}
}
\concept{whimsical generators}
