% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seazonlipsum.R
\name{seazon}
\alias{seazon}
\title{Generate Lorem Ipsum using the Seazon.org API}
\source{
\url{http://www.seazon.org/}
}
\usage{
seazon(
  latin = TRUE,
  start_with_lorem = TRUE,
  use_periods = TRUE,
  use_sentence_case = TRUE,
  p = TRUE,
  use_double_spaces = FALSE,
  random_numbers = FALSE,
  random_latin = TRUE,
  min_letters = 2L,
  max_letters = 9L,
  min_words = 45L,
  max_words = 85L,
  min_paras = 3L,
  max_paras = 4L
)
}
\arguments{
\item{latin}{Logical. If \code{TRUE}, generates Latin text. Otherwise generates random text. Defaults to \code{TRUE}.}

\item{start_with_lorem}{Logical. If \code{TRUE}, begins the text with 'Lorem Ipsum'. Defaults to \code{TRUE}.}

\item{use_periods}{Logical. If \code{TRUE}, uses periods. Defaults to \code{TRUE}.}

\item{use_sentence_case}{Logical. If \code{TRUE}, begins each sentence with an upper case character. Defaults to \code{TRUE}.}

\item{p}{Logical. If \code{TRUE}, wraps text in paragraph tags. Defaults to \code{TRUE}.}

\item{use_double_spaces}{Logical. If \code{TRUE}, puts double spaces after each sentence. Defaults to \code{FALSE}.}

\item{random_numbers}{Logical. If \code{TRUE}, adds random numbers if 'latin = FALSE'. Defaults to \code{TRUE}.}

\item{random_latin}{Logical. If \code{TRUE}, makes the random text latin-like if 'latin = FALSE'. Defaults to \code{TRUE}.}

\item{min_letters, max_letters}{Minimum/maximum numbers of letters in each word, when 'latin = FALSE'.}

\item{min_words, max_words}{Minimum/maximum paragraph word count.}

\item{min_paras, max_paras}{Minimum/maximum number of paragraphs.#'}
}
\description{
Generates random, customisable placeholder text using the Seazon API.
}
\seealso{
Other practical generators: 
\code{\link{corporatelorem}()},
\code{\link{loripsum}()},
\code{\link{markdownum}()}
}
\concept{practical generators}
