% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipsum.R
\docType{data}
\name{lipsum}
\alias{lipsum}
\title{The lipsum corpus}
\format{A data frame with 3189 rows and 4 variables:
\describe{
  \item{paragraph}{an integer id of the paragraph}
  \item{characters}{the number of characters in the sentence}
  \item{words}{the number of words in the sentence}
  \item{sentence}{the actual sentence}
}}
\description{
A data frame with sentences from the lipsum corpuse
}
\keyword{data}
