% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paragraph.R
\name{paragraph}
\alias{paragraph}
\title{return lipsum paragraphs}
\usage{
paragraph(id, words, n = 1, verbose = TRUE)
}
\arguments{
\item{id}{the id of the paragraph}

\item{words}{an optional prefered number of words. Only used when \code{ìd} is not given.}

\item{n}{the number of random paragraphs. Only used when \code{id} is not given.}

\item{verbose}{print the paragraphs}
}
\value{
an invisible vector of paragraphs
}
\description{
return lipsum paragraphs
}
\examples{
paragraph()
paragraph(n = 2)
paragraph(10)
paragraph(c(7, 1))
z <- paragraph(1, verbose = FALSE)
cat(z)
paragraph(words = 10, n = 2)
}
