% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paragraph.R
\name{paragraph}
\alias{paragraph}
\title{return lipsum paragraphs}
\usage{
paragraph(id, words, characters, n = 1)
}
\arguments{
\item{id}{the id of the paragraph}

\item{words}{an optional prefered number of words. Only used when \code{ìd} is not given.}

\item{characters}{an optional prefered number of characters. Only used when \code{ìd} and \code{words} are not given.}

\item{n}{the number of random paragraphs. Only used when \code{id} is not given.}
}
\value{
an string of paragraphs. The paragraphs are separated by a double newline character.
}
\description{
return lipsum paragraphs
}
\examples{
paragraph()
paragraph(n = 2)
paragraph(10)
paragraph(c(7, 1))
paragraph(words = 12, n = 2)
paragraph(characters = 90)
}
