% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paragraph.R
\name{paragraph}
\alias{paragraph}
\title{return lipsum paragraphs}
\usage{
paragraph(id, verbose = TRUE)
}
\arguments{
\item{id}{the id of the paragraph}

\item{verbose}{print the paragraphs}
}
\value{
an invisible vector of paragraphs
}
\description{
return lipsum paragraphs
}
\examples{
paragraph(10)
paragraph(c(7, 1))
z <- paragraph(1, verbose = FALSE)
cat(z)
}
