% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence.R
\name{sentence}
\alias{sentence}
\title{Return sentences from the lipsum corpus}
\usage{
sentence(id, n = 1, words, characters)
}
\arguments{
\item{id}{the number of the sentence}

\item{n}{the number of random sentences. Only used when \code{id} is not given.}

\item{words}{an optional prefered number of words. Only used when \code{ìd} is not given.}

\item{characters}{an optional prefered number of characters. Only used when \code{ìd} and \code{words} are not given.}
}
\value{
a string of sentences
}
\description{
A single paragraph with the sentences will be displayed
}
\examples{
sentence()
sentence(1)
sentence(2:3)
sentence(words = 5)
}
